layout/base/tests/test_bug465448.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="/tests/SimpleTest/test.css" type="text/css"?>

<window title="Bug 465448"
  onload="loaded()"
  xmlns:html="http://www.w3.org/1999/xhtml"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>      
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>      

<script><![CDATA[
SimpleTest.waitForExplicitFinish();
var loadedCalled = false;
var win = window.open("data:text/html,<body onload='window.opener.loaded()'><div style='height:200px; width:100px;'>", "_blank", "width=600,height=600");

function loaded() {
  if (!loadedCalled) {
    loadedCalled = true;
    return;
  }
  win.sizeToContent();
  win.sizeToContent();
  win.sizeToContent();
  win.sizeToContent();
  win.sizeToContent();
  win.sizeToContent();
  ok(win.innerWidth >= 100, "innerWidth");
  ok(win.innerHeight >= 200, "innerHeight");
  win.close();
  SimpleTest.finish();
}
]]></script>

<body xmlns="http://www.w3.org/1999/xhtml">
<p id="display">
</p>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
</body>

</window>
